requirements.txt
flax==0.10.6
rlds
tensorflow_datasets==4.9.9
--find-links https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
dm-reverb==0.14.0
ipython==7.34.0
matplotlib==3.7.2
jax[cuda12_pip]
orbax-checkpoint==0.4.4
